v12/backend/client/requirements.txt
grpcio
grpcio-tools
Flask
flask-cors
itsdangerous
Jinja2
MarkupSafe
requests
SimpleJSON